Add performance test as well without fi
